// Frontend/src/app/api/models/create-job-offer-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateJobOfferDto {
  dateLimite?: string;
  description: string;
  experienceMin: number;
  jobTitle: string;
  jobType: 'PER' | 'PATS' | 'CONTRACTUEL';
  niveauEtudeRequis: string;
  requiredSkills: string;
  typeContrat: 'CDD' | 'CDI' | 'STAGE';
}